Expire Policy and Withdraw Won Premiums
`expirePolicy` is a public endpoint that can be called by anyone after
the policy expired. It resolves the policy with zero payout,
distributing the premiums and unlocking the SCR.

`withdrawWonPremiums` this endpoint is restricted to used with role
WITHDRAW_WON_PREMIUMS_ROLE and withdraws the won pure premiums to the
treasury. Can be disabled revoking the WITHDRAW_WON_PREMIUMS_ROLE.