Immutable attributes
Use immutable attributes set in constructor for fields that don't change
in contract lifetime.

Update to latest OZ-updates library that supports constructorArgs.

Remove events Deposit and Withdraw to decrease size of PolicyPool (not
needed because ETokens anyway emit similar events as Transfer with
address(0) in `from` on Deposit and `to` on Withdraw.