Refactoring in EToken to use TimeScaled amount
Simplifies the code since we are not using the same library used in pool
loans.

Also saves gas since we reduced from 3 words (256 bit) to just 1.

Change in TimeScaled.ScaledAmount to use 128 bits in the scale
(previously was just 96 that tolerates less than 100x).

lastUpdate reduced to 32 bits, that should be fine for the next 80
years.