QSP-8 Fix compiler version with pragma
Requested by the auditors, in the core repository files we fix the
pragma to reflect this is the version that has been tested and audited
and this is the version that should be deployed.

At the same time, to simplify the life of developers using the
@ensuro/core NPM package, we unlock the pragma for the files included in
that package. This way, developers can use Ensuro contracts with other
solidity versions, at their own risk.