Loss Propagation Limits
Adds two new parameters to PremiumsAccount to be able to specify the
limits on the loans taken from the Junior and Senior eToken. By default
the loans are unbounded.

These parameters allow to define a hard limit on the propagation of
losses to the eTokens, mainly to the senior one.

This could have been implemented on eToken side, but that would require
a major refactoring, so we keep that for next versions.