Refactor PremiumsAccount - Repayments on-demand
Previously it was trying to do the repayment of the loans automatically
in each point of the code where a pure premium was won. This wasn't
effective because it's a waste of resources when the PA is not in debt
(most of the time). Also, when the PA is in debt this produces a lot of
micro-payments increasing the gas usage and also making the PA more
prone to take debt in the near term.

Now the debt repayment has been moved to a new public method. This can
be called at any time given enough funds available.

Funds available are calculated as (surplus - _maxDeficit), so the PA
will take as much deficit as needed to repay the debt.

Also in the previous version the debt was only repaid with won pure
premiums. Given this change, it can be paid with pure premiums comming
from new business. A good risk management must be made to avoid
postponing the bad performance of a growing portfolio to the future.