Refactoring: tokenInterestRate => scr earnings
Previously we were keeping the tokenInterestRate, computed as
`scrInterestRate * scr / totalSupply`. But this was complex and
inaccurate, since the tokenInterestRate was updated on every scr change,
but the totalSupply grows continously.

Now, we just store the scr and scrInterestRate, and on every change we
project the scale this way:

If `ts = a * s` (ts=totalSupply, s=scale, a=scaled amount or shares,
compatible with the values stored in _balances)

Then, we reflect the SCR earnings this way:

`ts + earnings = a * (s + earnings / a)`, then the `new_s = (s + earnings / a)`.

Also, I changed the ETKLib code so all the functions are views and they
return the modified struct, making easier to check in the EToken code
when the `_scr` or `_tsScaled` is changed.

The ScaledAmount that grows at a given interest rate was kept, because
internalLoans still work that way.