Refactoring in RiskModule: extract IUnderwriter
Refactoring of RiskModule, now everything related to the pricing of the
policy is delegated to another contract, the "Underwriter", that
receives the input and returns the information required to create the
policy.

Now the RiskModule is a concrete class, no longer an abstract class, and
should be used directly.

For the Underwriter contracts we might have different implementations,
depending if the input has all the information to create the policy, or
some parameters are stored on-chain.

Work in progress, still I have to implement another Underwriter (full
signed) and fix many tests.