EIP-2612 (ERC20 signed approval) + operate on behalf
Adds EIP-2612 to eTokens. Also, adds a new method `depositWithPermit`
that supports ERC-20 signed approval (EIP-2612) for doing gasless
approvals on the USDC.

Also, allows deposits and withdrawals on behalf of other users. For
deposit, the caller can mint the tokens for someone else.

For withdrawal, the caller might not be the owner of the tokens but
instead have allowance. And can decide to send the resulting USDC to a
receiver account.

Work in progress, all tests broken.